ai/requirements.txt
langchain==0.1.16
langchain-community==0.0.36
langchain-experimental==0.0.57
langchain-openai==0.1.4
langchain-pinecone==0.1.0
python-dotenv==1.0.1